+++ dataserver/spatial/transform.h
@@ -46,6 +46,7 @@ struct transform : is_static {
     static Meters STLength(spatial_point const * first, spatial_point const * end);
     static track_closest_point_t STClosestpoint(spatial_point const * first, spatial_point const * end, spatial_point const & where);
     static Degree course_between_points(spatial_point const & p1, spatial_point const & p2);
+    static Degree angle_between_points(spatial_point const & p1, spatial_point const & p2, spatial_point const & p3);
 };
 
 struct transform_t : is_static {